Se segmentan las pruebas y se verifica que funcionen
